Handle edge cases (#292)
* Handle cover.source.png and cover.source.tif

Artwork::CreateFromFilesystem expects a JPEG, so convert the PNG or TIFF to JPEG if necessary.

* Add the repo dir to print statements for debugging

When processing many ebooks in a loop, it can be hard to isolate the one that caused the error without the repo dir on the error line

* Add artists found from cover.source.{png,tif}

I missed these artists earlier because the code didn't handle cover.source.png
and cover.source.tif artwork.

* Handle content.opf that have artist-1

Instead of <dc:contributor id="artist">, some content.opf files have "artist-1"
and "artist-2" (and very rarely "artist-3"). Taking "artist-1" isn't ideal, but
is usually correct. Sometimes "artist-2" is for artwork after the cover.
Sometimes it indicates that two artists worked on the cover, which we aren't
set up to handle.

* Add artists found from artist-1 in content.opf